implemented an if statement to render library when a user logs in
